Moved database subscription into gridstore, created uninitialization function instead of allowing the isInitialized boolean to be changed outside the store.
