+++ src/App.vue
@@ -5,7 +5,6 @@ import { useGridStore } from './stores/gridStore'
 import {
   insertCell,
   updateCell,
-  subscribeToCellsChange,
 } from './services/supabase'
 import Cell from './components/Cell.vue'
 import {
@@ -15,20 +14,17 @@ import {
   DEFAULT_COLOR
 } from '@/utils/grid.js'
 
-let cellsSubscription
 const gridStore = useGridStore()
 
 const size = '25px'
 
 onMounted(async () => {
-  cellsSubscription = subscribeToCellsChange(onInsert, onUpdate, onDelete)
   await gridStore.initialize()
 })
 
 onUnmounted(() => {
   console.log('unmount')
-  gridStore.isInitialized = false
-  cellsSubscription.unsubscribe()
+  gridStore.uninitialize()
 })
 
 const cssGridColumnTemplate = computed(() => {
@@ -38,24 +34,6 @@ const cssGridColumnTemplate = computed(() => {
     return 'unset'
   }
 })
-
-function onInsert(payload) {
-  if (!gridStore.isInitialized) { return }
-  console.log('onInsert()', payload)
-  gridStore.insert(payload.new)
-}
-
-function onUpdate(payload) {
-  if (!gridStore.isInitialized) { return }
-  console.log('onUpdate()', payload)
-  gridStore.update(payload.new)
-}
-
-function onDelete(payload) {
-  if (!gridStore.isInitialized) { return }
-  console.log('onDelete()', payload)
-  gridStore.remove(payload.old.id)
-}
 
 function colorFromStore(index) {
   if (!gridStore.isInitialized) { return }
